% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_types.R
\name{character_long}
\alias{character_long}
\title{SQL classes without a directly analogous R class
character_long <=> VARCHAR(30000)}
\usage{
character_long()
}
\description{
SQL classes without a directly analogous R class
character_long <=> VARCHAR(30000)
}

