% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{A function to create a query to create keyed data.tables into a MySQL database}
\usage{
create(model, verbose = TRUE)
}
\description{
A function to create a query to create keyed data.tables into a MySQL database
}

