% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_types.R
\name{date_time}
\alias{date_time}
\title{POXISt <=> DATETIME
An empty POSIXt is otherwise classed as a character}
\usage{
date_time()
}
\description{
POXISt <=> DATETIME
An empty POSIXt is otherwise classed as a character
}

