% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetKey.R
\name{dbGetKey}
\alias{dbGetKey}
\title{dbGetKey}
\usage{
dbGetKey(con, name)
}
\arguments{
\item{con}{A database connection}

\item{name}{The name of the MySQL table}
}
\description{
Get the primary key from a table
}
