% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSyncTableR.R
\name{dbSyncTable}
\alias{dbSyncTable}
\title{A function to create a query to create keyed data.tables into a MySQL database}
\usage{
dbSyncTable(con, model, ...)
}
\description{
A function to create a query to create keyed data.tables into a MySQL database
}

