% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbUpdateTable.R
\name{dbUpdateTable}
\alias{dbUpdateTable}
\title{Update rows of a keyed table via DELETE + INSERT}
\usage{
dbUpdateTable(con, name, dt, ...)
}
\arguments{
\item{con}{A MySQL connection}

\item{name}{A MySQL table name}

\item{dt}{A keyed data.table with data to update in \code{name}}
}
\description{
Update rows of a keyed table via DELETE + INSERT
}

