% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql_type.R
\name{get_sql_type}
\alias{get_sql_type}
\title{get_sql_type}
\usage{
get_sql_type(type, dt = hash_datatypes)
}
\arguments{
\item{type}{The R data type to switch to a MySQL data type}

\item{dt}{The hash table of types}
}
\description{
Internal switch between R types and MySQL types
}

