% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%format\%}
\alias{\%format\%}
\title{%format% infix to format strings via named paramters}
\usage{
fmt \%format\% list
}
\arguments{
\item{fmt}{A string to format}

\item{list}{A list of named parameters}
}
\description{
%format% infix to format strings via named paramters
}
\examples{
parameters <- list(label = "months", april = 4, may = 5, june = 6)
"\%(label)s: \%(april)d \%(may)d \%(june)d" \%format\% parameters

}

