% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql_type.R
\docType{data}
\name{hash_datatypes}
\alias{hash_datatypes}
\title{Current implemented datatype handling}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 4 rows and 2 columns.}
\usage{
hash_datatypes
}
\description{
Current implemented datatype handling
}
\keyword{datasets}

