% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_types.R
\name{long_text}
\alias{long_text}
\title{SQL classes without a directly analogous R class
Long text <=> VARCHAR(30000)}
\usage{
long_text()
}
\description{
SQL classes without a directly analogous R class
Long text <=> VARCHAR(30000)
}

