% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbSyncTableR.R
\name{create}
\alias{create}
\title{Create}
\usage{
create(model, verbose = TRUE, ...)
}
\arguments{
\item{model}{A \code{data.table} model - by convention is prefaced with \code{model_}}

\item{verbose}{Boolean to control printing of created statement}

\item{...}{Carry parameters forward from dbSyncTable}
}
\description{
Create a query to which when executed will create keyed data.tables in a MySQL database
}

